[SMA-280] Move evoleq/util to evoleq/ktorx step 1
