+++ solawi-bid-backend/src/main/kotlin/org/evoleq/ktorx/contextual.kt
@@ -1,4 +1,4 @@
-package org.evoleq.util
+package org.evoleq.ktorx
 
 import java.util.UUID
