+++ solawi-bid-backend/src/main/kotlin/org/evoleq/ktorx/io.kt
@@ -1,6 +1,5 @@
-package org.evoleq.util
+package org.evoleq.ktorx
 
-// import org.solyton.solawi.bid.application.plugin.AuthenticationHolder.Companion.jwtPrincipal
 import io.ktor.http.*
 import io.ktor.server.auth.*
 import io.ktor.server.auth.jwt.*
@@ -84,68 +83,8 @@ suspend inline fun <reified T : Any>  Receive(d: T): Action<Result<T>> = ApiActi
 
 @KtorDsl
 @Suppress("FunctionName")
-suspend inline fun <reified T : Any>  Respond(): KlAction<Result<T>, Unit> = {result ->
-    ApiAction { call->
-        when(result){
-            is Result.Success<T> -> try{
-                call.respond(
-                    HttpStatusCode.OK,
-                    Json.encodeToString( ResultSerializer(), result)
-                )
-            } catch (exception: Exception) {
-                call.respond(
-                    HttpStatusCode.InternalServerError,
-                    exception.message?: "Serialization Error"
-                )
-            }
-
-            is Result.Failure.Message ->  call.respond(
-                HttpStatusCode.InternalServerError,
-                result
-            )
-
-            is Result.Failure.Exception -> with(result.transform()){
-                call.respond(first, second)
-            }
-        } x call
-    }
-}
-
-@KtorDsl
-@Suppress("FunctionName")
 fun <S : Any, T: Any> Transform(f: (S)-> T): KlAction<Result<S>, Result<T>> =
     {result: Result<S> -> Action { base -> result map f x base } }
-
-fun Result.Failure.Exception.transform(): Pair<HttpStatusCode, Result.Failure.Message> =
-    when(this.value) {
-        // Authentication
-        is AuthenticationException.InvalidOrExpiredToken -> HttpStatusCode.Unauthorized
-
-        // BidRound
-        is BidRoundException.RoundNotStarted -> HttpStatusCode.Conflict
-        is BidRoundException.NoSuchRound,
-        is BidRoundException.NoSuchRoundState,
-        is BidRoundException.NoSuchAuction -> HttpStatusCode.NotFound
-        is BidRoundException.UnregisteredBidder,
-        is BidRoundException.RegisteredBidderNotPartOfTheAuction,
-        is BidRoundException.AuctionAccepted,
-        is BidRoundException.LinkNotPresent, -> HttpStatusCode.Forbidden
-        is BidRoundException.IllegalNumberOfParts -> HttpStatusCode.BadRequest
-        is BidRoundException.MissingBidderDetails -> HttpStatusCode.NotFound
-        //
-        //User
-        is UserManagementException.UserDoesNotExist -> HttpStatusCode.Unauthorized
-        is UserManagementException.WrongCredentials -> HttpStatusCode.Unauthorized
-
-        // Permission
-        is PermissionException.AccessDenied -> HttpStatusCode.Forbidden
-
-        // RoundState
-        is RoundStateException.IllegalTransition -> HttpStatusCode.BadRequest
-        is RoundStateException.IllegalRoundState -> HttpStatusCode.BadRequest
-
-        else -> HttpStatusCode.InternalServerError
-    } x this.value.toMessage()
 
 fun Throwable.toMessage(): Result.Failure.Message = Result.Failure.Message(message?: NO_MESSAGE_PROVIDED)
