+++ solawi-bid-backend/src/main/kotlin/org/solyton/solawi/bid/module/authentication/action/is-logged-in.kt
@@ -5,7 +5,7 @@ import io.ktor.util.*
 import org.evoleq.exposedx.transaction.resultTransaction
 import org.evoleq.ktorx.result.Result
 import org.evoleq.math.x
-import org.evoleq.util.Action
+import org.evoleq.ktorx.Action
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
 import org.jetbrains.exposed.sql.deleteWhere
 import org.joda.time.DateTime
