+++ solawi-bid-backend/src/main/kotlin/org/solyton/solawi/bid/module/bid/action/api/bid.kt
@@ -2,8 +2,8 @@ package org.solyton.solawi.bid.module.bid.action.api
 
 import io.ktor.server.request.*
 import org.evoleq.math.x
-import org.evoleq.util.Action
-import org.evoleq.util.ApiAction
+import org.evoleq.ktorx.Action
+import org.evoleq.ktorx.ApiAction
 import org.solyton.solawi.bid.module.bid.data.api.Bid
 
 val ReceiveBid: Action<Bid> = ApiAction {
