+++ solawi-bid-backend/src/main/kotlin/org/solyton/solawi/bid/module/bid/action/db/auction.kt
@@ -8,8 +8,8 @@ import org.evoleq.ktorx.result.bindSuspend
 import org.evoleq.ktorx.result.map
 import org.evoleq.math.MathDsl
 import org.evoleq.math.x
-import org.evoleq.util.DbAction
-import org.evoleq.util.KlAction
+import org.evoleq.ktorx.DbAction
+import org.evoleq.ktorx.KlAction
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.inList
 import org.jetbrains.exposed.sql.Transaction
 import org.jetbrains.exposed.sql.deleteWhere
