+++ solawi-bid-backend/src/main/kotlin/org/solyton/solawi/bid/module/bid/action/db/bid.kt
@@ -5,8 +5,8 @@ import org.evoleq.ktorx.result.Result
 import org.evoleq.ktorx.result.bindSuspend
 import org.evoleq.math.MathDsl
 import org.evoleq.math.x
-import org.evoleq.util.DbAction
-import org.evoleq.util.KlAction
+import org.evoleq.ktorx.DbAction
+import org.evoleq.ktorx.KlAction
 import org.jetbrains.exposed.sql.Transaction
 import org.jetbrains.exposed.sql.and
 import org.solyton.solawi.bid.module.bid.data.addBidInfo
