+++ solawi-bid-backend/src/main/kotlin/org/solyton/solawi/bid/module/permission/action/db/is-granted.kt
@@ -4,9 +4,9 @@ import org.evoleq.exposedx.transaction.resultTransaction
 import org.evoleq.ktorx.result.*
 import org.evoleq.math.MathDsl
 import org.evoleq.math.x
-import org.evoleq.util.Contextual
-import org.evoleq.util.DbAction
-import org.evoleq.util.KlAction
+import org.evoleq.ktorx.Contextual
+import org.evoleq.ktorx.DbAction
+import org.evoleq.ktorx.KlAction
 import org.jetbrains.exposed.sql.*
 import org.solyton.solawi.bid.module.permission.schema.Contexts
 import org.solyton.solawi.bid.module.permission.schema.ContextsTable
