+++ solawi-bid-backend/src/main/kotlin/org/solyton/solawi/bid/module/permission/action/db/user-permissions.kt
@@ -1,14 +1,13 @@
 package org.solyton.solawi.bid.module.permission.action.db
 
-import io.ktor.client.plugins.convertLongTimeoutToIntWithInfiniteAsZero
 import org.evoleq.exposedx.transaction.resultTransaction
 import org.evoleq.ktorx.result.Result
 import org.evoleq.ktorx.result.bindSuspend
 import org.evoleq.math.MathDsl
 import org.evoleq.math.x
-import org.evoleq.util.Contextual
-import org.evoleq.util.DbAction
-import org.evoleq.util.KlAction
+import org.evoleq.ktorx.Contextual
+import org.evoleq.ktorx.DbAction
+import org.evoleq.ktorx.KlAction
 import org.jetbrains.exposed.sql.Transaction
 import org.jetbrains.exposed.sql.and
 import org.jetbrains.exposed.sql.selectAll
